factors logger function into separate source
